ptt-box/pyproject.toml
[project]
name = "ptt-box"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "setuptools<70",  # pkg_resources required by ctranslate2
    "numpy==1.26.4",
    "sounddevice>=0.5.3",
    "faster-whisper==1.0.2",
    "watchdog>=4.0.0",
    "onnxruntime==1.17.3",
    "ctranslate2==4.2.1",
    "tokenizers==0.19.1",
    "python-dotenv>=1.0.0",
    # AI Assistant dependencies
    "openai>=1.0.0",
    "edge-tts>=6.1.0",
    "fastmcp>=2.0.0",
    "aiohttp>=3.9.0",
]
